x86-64: Set up boot GDT
To enter 64-bit long mode we need to add a 64-bit code segment to the
global descriptor table. But according to the Multiboot specification
the value in GDTR may be invalid -- the GDT set up by the boot loader
may have been overwritten. We need to set up our own GDT.

The GDT is an array of segment descriptors, each of which defines the
base address, limit (size), and flags associated with a segment. In
addition to the required "null" segment, we define two "flat" 32-bit
segments (an executable code segment and a read-write data segment)
spanning the 4 GiB linear address space.

After loading GDTR with the size and address of our GDT, we load all
segment registers with selectors pointing to our new segments.